refactor(dashboard): extract EditCompatibleNodeModal — #3501 Phase 1b
Continues the strangler-fig decomposition. Extracts EditCompatibleNodeModal
(+ its node/props types) into components/modals/, and moves the shared
CC_COMPATIBLE_DEFAULT_CHAT_PATH constant into a leaf providerDetailConstants.ts
so the page client and the modal both import it without a circular dependency
(check:cycles clean). Removes two dangling section comments left by Phase 1a.

ProviderDetailPageClient.tsx: 10,719 -> 10,435 LOC. Behavior unchanged: Phase 0
smoke test still green + a new standalone modal render test. typecheck:core,
lint, prettier, check:cycles all clean; file-size ratchet lowered.

Part of #3501. Co-authored-by: oyi77 <oyi77@users.noreply.github.com>
