+++ src/app/(dashboard)/dashboard/providers/[id]/ProviderDetailPageClient.tsx
@@ -99,6 +99,8 @@ import { ImportCodexAuthModal, ApplyCodexAuthModal } from "./components/modals/I
 import { ImportClaudeAuthModal, ApplyClaudeAuthModal } from "./components/modals/ImportClaudeAuthModal";
 import { ImportGeminiAuthModal, ApplyGeminiAuthModal } from "./components/modals/ImportGeminiAuthModal";
 
+import EditCompatibleNodeModal from "./components/modals/EditCompatibleNodeModal";
+import { CC_COMPATIBLE_DEFAULT_CHAT_PATH } from "./providerDetailConstants";
 type CompatByProtocolMap = Partial<
   Record<
     ModelCompatProtocolKey,
@@ -905,27 +907,6 @@ interface EditConnectionModalProps {
   onSave: (data: unknown) => Promise<void | unknown>;
   onClose: () => void;
 }
-
-interface EditCompatibleNodeModalNode {
-  id?: string;
-  name?: string;
-  prefix?: string;
-  apiType?: string;
-  baseUrl?: string;
-  chatPath?: string;
-  modelsPath?: string;
-}
-
-interface EditCompatibleNodeModalProps {
-  isOpen: boolean;
-  node: EditCompatibleNodeModalNode | null;
-  onSave: (data: unknown) => Promise<void>;
-  onClose: () => void;
-  isAnthropic?: boolean;
-  isCcCompatible?: boolean;
-}
-
-const CC_COMPATIBLE_DEFAULT_CHAT_PATH = "/v1/messages?beta=true";
 const CODEX_REASONING_STRENGTH_OPTIONS = [
   { value: "none", label: "None" },
   { value: "low", label: "Low" },
@@ -9347,8 +9328,6 @@ function AddApiKeyModal({
   );
 }
 
-// ──── ImportCodexAuthModal ────────────────────────────────────────────────────
-
 function normalizeAndValidateHttpBaseUrl(rawValue, fallbackUrl) {
   const value = (typeof rawValue === "string" ? rawValue.trim() : "") || fallbackUrl;
   try {
@@ -10453,267 +10432,3 @@ function EditConnectionModal({ isOpen, connection, onSave, onClose }: EditConnec
     </Modal>
   );
 }
-
-function EditCompatibleNodeModal({
-  isOpen,
-  node,
-  onSave,
-  onClose,
-  isAnthropic,
-  isCcCompatible,
-}: EditCompatibleNodeModalProps) {
-  const t = useTranslations("providers");
-  const [formData, setFormData] = useState({
-    name: "",
-    prefix: "",
-    apiType: "chat",
-    baseUrl: "https://api.openai.com/v1",
-    chatPath: "",
-    modelsPath: "",
-  });
-  const [saving, setSaving] = useState(false);
-  const [checkKey, setCheckKey] = useState("");
-  const [validating, setValidating] = useState(false);
-  const [validationResult, setValidationResult] = useState(null);
-  const [showAdvanced, setShowAdvanced] = useState(false);
-
-  useEffect(() => {
-    if (node) {
-      setFormData({
-        name: node.name || "",
-        prefix: node.prefix || "",
-        apiType: node.apiType || "chat",
-        baseUrl:
-          node.baseUrl ||
-          (isCcCompatible
-            ? "https://api.anthropic.com"
-            : isAnthropic
-              ? "https://api.anthropic.com/v1"
-              : "https://api.openai.com/v1"),
-        chatPath: node.chatPath || (isCcCompatible ? CC_COMPATIBLE_DEFAULT_CHAT_PATH : ""),
-        modelsPath: isCcCompatible ? "" : node.modelsPath || "",
-      });
-      setShowAdvanced(
-        !!(
-          node.chatPath ||
-          (!isCcCompatible && node.modelsPath) ||
-          (isCcCompatible && !node.chatPath)
-        )
-      );
-    }
-  }, [node, isAnthropic, isCcCompatible]);
-
-  const apiTypeOptions = [
-    { value: "chat", label: t("chatCompletions") },
-    { value: "responses", label: t("responsesApi") },
-    { value: "embeddings", label: t("embeddings") },
-    { value: "audio-transcriptions", label: t("audioTranscriptions") },
-    { value: "audio-speech", label: t("audioSpeech") },
-    { value: "images-generations", label: t("imagesGenerations") },
-  ];
-
-  const handleSubmit = async () => {
-    if (!formData.name.trim() || !formData.prefix.trim() || !formData.baseUrl.trim()) return;
-    setSaving(true);
-    try {
-      const payload: any = {
-        name: formData.name,
-        prefix: formData.prefix,
-        baseUrl: formData.baseUrl,
-        chatPath: formData.chatPath || (isCcCompatible ? CC_COMPATIBLE_DEFAULT_CHAT_PATH : ""),
-        modelsPath: isCcCompatible ? "" : formData.modelsPath,
-      };
-      if (!isAnthropic) {
-        payload.apiType = formData.apiType;
-      }
-      await onSave(payload);
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const handleValidate = async () => {
-    setValidating(true);
-    try {
-      const res = await fetch("/api/provider-nodes/validate", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          baseUrl: formData.baseUrl,
-          apiKey: checkKey,
-          type: isAnthropic ? "anthropic-compatible" : "openai-compatible",
-          compatMode: isCcCompatible ? "cc" : undefined,
-          chatPath: formData.chatPath || (isCcCompatible ? CC_COMPATIBLE_DEFAULT_CHAT_PATH : ""),
-          modelsPath: isCcCompatible ? "" : formData.modelsPath,
-        }),
-      });
-      const data = await res.json();
-      setValidationResult(data.valid ? "success" : "failed");
-    } catch {
-      setValidationResult("failed");
-    } finally {
-      setValidating(false);
-    }
-  };
-
-  if (!node) return null;
-
-  return (
-    <Modal
-      isOpen={isOpen}
-      title={
-        isCcCompatible
-          ? t("ccCompatibleDetailsTitle")
-          : t("editCompatibleTitle", { type: isAnthropic ? t("anthropic") : t("openai") })
-      }
-      onClose={onClose}
-    >
-      <div className="flex flex-col gap-4">
-        {isCcCompatible && (
-          <div className="rounded-lg border border-amber-500/25 bg-amber-500/10 px-3 py-2 text-sm text-text-muted">
-            <div className="flex items-start gap-2">
-              <span className="material-symbols-outlined mt-0.5 text-[18px] text-amber-500">
-                warning
-              </span>
-              <p>{t("ccCompatibleValidationHint")}</p>
-            </div>
-          </div>
-        )}
-        <Input
-          label={t("nameLabel")}
-          value={formData.name}
-          onChange={(e) => setFormData({ ...formData, name: e.target.value })}
-          placeholder={
-            isCcCompatible
-              ? t("ccCompatibleNamePlaceholder")
-              : t("compatibleProdPlaceholder", {
-                  type: isAnthropic ? t("anthropic") : t("openai"),
-                })
-          }
-          hint={isCcCompatible ? t("ccCompatibleNameHint") : t("nameHint")}
-        />
-        <Input
-          label={t("prefixLabel")}
-          value={formData.prefix}
-          onChange={(e) => setFormData({ ...formData, prefix: e.target.value })}
-          placeholder={
-            isCcCompatible
-              ? t("ccCompatiblePrefixPlaceholder")
-              : isAnthropic
-                ? t("anthropicPrefixPlaceholder")
-                : t("openaiPrefixPlaceholder")
-          }
-          hint={isCcCompatible ? t("ccCompatiblePrefixHint") : t("prefixHint")}
-        />
-        {!isAnthropic && (
-          <Select
-            label={t("apiTypeLabel")}
-            options={apiTypeOptions}
-            value={formData.apiType}
-            onChange={(e) => setFormData({ ...formData, apiType: e.target.value })}
-          />
-        )}
-        <Input
-          label={t("baseUrlLabel")}
-          value={formData.baseUrl}
-          onChange={(e) => setFormData({ ...formData, baseUrl: e.target.value })}
-          placeholder={
-            isCcCompatible
-              ? t("ccCompatibleBaseUrlPlaceholder")
-              : isAnthropic
-                ? t("anthropicBaseUrlPlaceholder")
-                : t("openaiBaseUrlPlaceholder")
-          }
-          hint={
-            isCcCompatible
-              ? t("ccCompatibleBaseUrlHint")
-              : t("compatibleBaseUrlHint", {
-                  type: isAnthropic ? t("anthropic") : t("openai"),
-                })
-          }
-        />
-        <button
-          type="button"
-          className="text-sm text-text-muted hover:text-text-primary flex items-center gap-1"
-          onClick={() => setShowAdvanced(!showAdvanced)}
-          aria-expanded={showAdvanced}
-          aria-controls="advanced-settings"
-        >
-          <span
-            className={`transition-transform ${showAdvanced ? "rotate-90" : ""}`}
-            aria-hidden="true"
-          >
-            ▶
-          </span>
-          {t("advancedSettings")}
-        </button>
-        {showAdvanced && (
-          <div id="advanced-settings" className="flex flex-col gap-3 pl-2 border-l-2 border-border">
-            <Input
-              label={t("chatPathLabel")}
-              value={formData.chatPath}
-              onChange={(e) => setFormData({ ...formData, chatPath: e.target.value })}
-              placeholder={
-                isCcCompatible
-                  ? CC_COMPATIBLE_DEFAULT_CHAT_PATH
-                  : isAnthropic
-                    ? "/messages"
-                    : t("chatPathPlaceholder")
-              }
-              hint={isCcCompatible ? t("ccCompatibleChatPathHint") : t("chatPathHint")}
-            />
-            {!isCcCompatible && (
-              <Input
-                label={t("modelsPathLabel")}
-                value={formData.modelsPath}
-                onChange={(e) => setFormData({ ...formData, modelsPath: e.target.value })}
-                placeholder={t("modelsPathPlaceholder")}
-                hint={t("modelsPathHint")}
-              />
-            )}
-          </div>
-        )}
-        <div className="flex gap-2">
-          <Input
-            label={t("apiKeyForCheck")}
-            type="password"
-            value={checkKey}
-            onChange={(e) => setCheckKey(e.target.value)}
-            className="flex-1"
-          />
-          <div className="pt-6">
-            <Button
-              onClick={handleValidate}
-              disabled={!checkKey || validating || !formData.baseUrl.trim()}
-              variant="secondary"
-            >
-              {validating ? t("checking") : t("check")}
-            </Button>
-          </div>
-        </div>
-        {validationResult && (
-          <Badge variant={validationResult === "success" ? "success" : "error"}>
-            {validationResult === "success" ? t("valid") : t("invalid")}
-          </Badge>
-        )}
-        <div className="flex gap-2">
-          <Button
-            onClick={handleSubmit}
-            fullWidth
-            disabled={
-              !formData.name.trim() || !formData.prefix.trim() || !formData.baseUrl.trim() || saving
-            }
-          >
-            {saving ? t("saving") : t("save")}
-          </Button>
-          <Button onClick={onClose} variant="ghost" fullWidth>
-            {t("cancel")}
-          </Button>
-        </div>
-      </div>
-    </Modal>
-  );
-}
-
-// ──── ImportGeminiAuthModal ────────────────────────────────────────────────────
-
